Add ability to report on rejection errors
FA's file spec for error repoting is _terible_, but I think I've got
something that is somewhat useful here. Essentially, if there are N
rejections then the last N lines of the error log are for those
(probably). Pair them together, select the ones we care about, and email
someone about it.